feat: 식물 등록 기능 개발 중

+++ src/main/java/com/example/plantappbackend/controller/WateringRecordController.java
@@ -21,7 +21,7 @@ public class WateringRecordController {
     }
 
     @GetMapping("/{plantId}")
-    public ResponseEntity<List<WateringRecord>> getRecords(@PathVariable Long plantId) {
+    public ResponseEntity<List<WateringRecord>> getRecords(@PathVariable int plantId) {
         return ResponseEntity.ok(wateringRecordService.getWateringRecords(plantId));
     }
 }
